add delet alert

+++ src/pages/Customers.tsx
@@ -1,231 +1,3 @@
-// import { ReactElement, useState } from "react";
-// import AdminSidebar from "../components/AdminSidebar"
-// import TableHOC from "../components/TableHOC"
-// import { createColumnHelper } from "@tanstack/react-table";
-// import { FaTrash } from "react-icons/fa";
-
-// interface DataType {
-//   avatar: ReactElement;
-//   name: string;
-//   email: string;
-//   gender: string;
-//   role: string;
-//   action: ReactElement;
-// }
-
-// const columnHelper = createColumnHelper<DataType>();
-//   const columns = [
-//     columnHelper.accessor("avatar", {
-//       header: () => 'Avatar',
-//       cell: (info) => info.getValue(),
-//       footer: (info) => info.column.id,
-//     }),
-//     columnHelper.accessor("name", {
-//       header: () => 'Name',
-//       cell: (info) => info.getValue(),
-//       footer: (info) => info.column.id,
-//     }),
-//     columnHelper.accessor("email", {
-//       header: () => "Email",
-//       cell: (info) => info.renderValue(),
-//       footer: (info) => info.column.id,
-//     }),
-//     columnHelper.accessor("role", {
-//       header: () => 'Role',
-//       cell: (info) => info.getValue(),
-//       footer: (info) => info.column.id,
-//     }),
-//     columnHelper.accessor("action", {
-//       header: () => 'Action',
-//       cell: (info) => info.getValue(),
-//       footer: (info) => info.column.id,
-//     })
-//   ];
-
-//   const img = "https://randomuser.me/api/portraits/women/54.jpg";
-// const img2 = "https://randomuser.me/api/portraits/women/50.jpg";
-
-//   const arr: DataType[] = [
-//     {
-//       avatar: (
-//         <img
-//           style={{
-//             borderRadius: "50%",
-//           }}
-//           src={img}
-//           alt="Shoes"
-//         />
-//       ),
-//       name: "Areesha Khan",
-//       email: "areesha.ak30@gmail.com",
-//       gender: "female",
-//       role: "user",
-//       action: (
-//         <button>
-//           <FaTrash />
-//         </button>
-//       ),
-//     },
-  
-//     {
-//       avatar: (
-//         <img
-//           style={{
-//             borderRadius: "50%",
-//           }}
-//           src={img2}
-//           alt="Shoes"
-//         />
-//       ),
-//       name: "Javeria Alam",
-//       email: "javeriaalam21@gmail.com",
-//       gender: "female",
-//       role: "user",
-//       action: (
-//         <button>
-//           <FaTrash />
-//         </button>
-//       ),
-//     }, {
-//       avatar: (
-//         <img
-//           style={{
-//             borderRadius: "50%",
-//           }}
-//           src={img2}
-//           alt="Shoes"
-//         />
-//       ),
-//       name: "Javeria Alam",
-//       email: "javeriaalam21@gmail.com",
-//       gender: "female",
-//       role: "user",
-//       action: (
-//         <button>
-//           <FaTrash />
-//         </button>
-//       ),
-//     }, {
-//       avatar: (
-//         <img
-//           style={{
-//             borderRadius: "50%",
-//           }}
-//           src={img2}
-//           alt="Shoes"
-//         />
-//       ),
-//       name: "Javeria Alam",
-//       email: "javeriaalam21@gmail.com",
-//       gender: "female",
-//       role: "user",
-//       action: (
-//         <button>
-//           <FaTrash />
-//         </button>
-//       ),
-//     }, {
-//       avatar: (
-//         <img
-//           style={{
-//             borderRadius: "50%",
-//           }}
-//           src={img2}
-//           alt="Shoes"
-//         />
-//       ),
-//       name: "Javeria Alam",
-//       email: "javeriaalam21@gmail.com",
-//       gender: "female",
-//       role: "user",
-//       action: (
-//         <button>
-//           <FaTrash />
-//         </button>
-//       ),
-//     }, {
-//       avatar: (
-//         <img
-//           style={{
-//             borderRadius: "50%",
-//           }}
-//           src={img2}
-//           alt="Shoes"
-//         />
-//       ),
-//       name: "Javeria Alam",
-//       email: "javeriaalam21@gmail.com",
-//       gender: "female",
-//       role: "user",
-//       action: (
-//         <button>
-//           <FaTrash />
-//         </button>
-//       ),
-//     }, {
-//       avatar: (
-//         <img
-//           style={{
-//             borderRadius: "50%",
-//           }}
-//           src={img2}
-//           alt="Shoes"
-//         />
-//       ),
-//       name: "Javeria Alam",
-//       email: "javeriaalam21@gmail.com",
-//       gender: "female",
-//       role: "user",
-//       action: (
-//         <button>
-//           <FaTrash />
-//         </button>
-//       ),
-//     }, {
-//       avatar: (
-//         <img
-//           style={{
-//             borderRadius: "50%",
-//           }}
-//           src={img2}
-//           alt="Shoes"
-//         />
-//       ),
-//       name: "Javeria Alam",
-//       email: "javeriaalam21@gmail.com",
-//       gender: "female",
-//       role: "user",
-//       action: (
-//         <button>
-//           <FaTrash />
-//         </button>
-//       ),
-//     },
-//   ];
-  
-
-// function Customers() {
-//   const [data,setArr] = useState<DataType[]>(arr)
-
-//     return (
-//       <div className="admin-container">
-//         <AdminSidebar />
-//         <main>
-//         {TableHOC<DataType>(
-//             columns,
-//             data,
-//             "dashboard-product-box",
-//             "Customers",
-//             true
-//           )}
-//         </main>
-//       </div>
-//     )
-//   }
-  
-//   export default Customers
-
-
 import { ReactElement, useState, useEffect } from "react";
 import AdminSidebar from "../components/AdminSidebar";
 import TableHOC from "../components/TableHOC";
@@ -278,7 +50,7 @@ function Customers() {
   useEffect(() => {
     axios.get('http://localhost:3002/users')
       .then(response => {
-        const users = response.data.map((user: any, index:any) => ({
+        const users = response.data.map((user: any, index: any) => ({
           avatar: (
             <img
               style={{
@@ -296,7 +68,7 @@ function Customers() {
           role: user.lastLoginDate,
           action: (
             <button
-              onClick={() => handleDelete(index,user._id)} 
+              onClick={() => handleDelete(index, user._id)} 
               style={{
                 width: "70px",
                 textDecoration: 'none',
@@ -315,24 +87,27 @@ function Customers() {
       .catch(error => {
         console.error('Error fetching users:', error);
         setLoading(false);
-
       });
   }, []);
 
   const handleDelete = (index: number, userId: string) => {
+    const confirmed = window.confirm('Are you sure you want to delete this user?');
+    if (!confirmed) {
+      return;
+    }
+
     axios.delete(`http://localhost:3002/users/${userId}`)
-    .then(() => {
-      setData(prevData => {
-        const newData = [...prevData];
-        newData.splice(index, 1);
-        return newData;
+      .then(() => {
+        setData(prevData => {
+          const newData = [...prevData];
+          newData.splice(index, 1);
+          return newData;
+        });
+      })
+      .catch(error => {
+        console.error('Error deleting user:', error);
       });
-    })
-    .catch(error => {
-      console.error('Error deleting product:', error);
-    });
   };
-  
 
   return (
     <div className="admin-container">
